raspberry/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
arandr==0.1.10
astroid==2.5.1
asttokens==2.0.4
asyncio==3.4.3
automationhat==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
distro==1.5.0
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
exceptiongroup==1.1.3
ExplorerHAT==0.4.2
fastapi==0.104.0
Flask==1.1.2
fourletterphat==0.1.0
gpiozero==1.6.2
h11==0.14.0
html5lib==1.1
httptools==0.6.1
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
lazy-object-proxy==0.0.0
logilab-common==1.8.1
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mypy==0.812
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.0
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
psutil==5.8.0
pycairo==1.16.2
pycups==2.0.1
pydantic==2.4.2
pydantic-core==2.10.1
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1
python-dotenv==1.0.0
python-prctl==1.7
PyYAML==6.0.1
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.12.1
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.4.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
sniffio==1.3.0
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
starlette==0.27.0
thonny==4.0.1
toml==0.10.1
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing-extensions==4.8.0
unicornhathd==0.0.4
urllib3==1.26.5
uvicorn==0.23.2
uvloop==0.18.0
v4l2-python3==0.3.2
watchfiles==0.21.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==1.0.1
wrapt==1.12.1
